requirements.txt
# === Core Dependencies ===
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1

# === AI & Embeddings ===
openai==1.30.1

# === PDF + Web Scraping ===
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.31.0

# === Email Handling ===
imap-tools==1.9.0

# === Database (SQLAlchemy + SQLite) ===
SQLAlchemy==2.0.30

# === Utilities ===
numpy==1.26.4
tqdm==4.66.4

# === Optional (Markdown to HTML) ===
markdown2==2.4.13
